[FS-B-7] Integration Tests

+++ FileService/src/FileService.Application/S3/IChunkSizeCalculator.cs
@@ -5,5 +5,5 @@ namespace FileService.Application.S3;
 
 public interface IChunkSizeCalculator
 {
-    public (int TotalChunks, long ChunkSize) Calculate(long fileSize);
+    public (int TotalChunks, int ChunkSize) Calculate(long fileSize);
 }
